Move InventoryFactory to convert
